requirements.txt
streamlit>=1.37
httpx>=0.27
openai>=1.30
duckdb>=1.0
pandas>=2.2
plotly>=5.22
numpy>=1.26
scikit-learn>=1.4
hdbscan>=0.8
umap-learn>=0.5
bibtexparser>=1.4
python-docx>=1.1
openpyxl>=3.1
pyvis>=0.3
